fix access to pages

+++ src/AppBundle/Controller/Cinema_hallController.php
@@ -22,13 +22,20 @@ class Cinema_hallController extends Controller
      */
     public function indexAction()
     {
-        $em = $this->getDoctrine()->getManager();
+        $loggedUserRole = $this->get('security.token_storage')->getToken()->getUser();
+        $hasAccess = $this->hasAccess($loggedUserRole);
 
-        $cinema_halls = $em->getRepository('AppBundle:Cinema_hall')->findAll();
+        if($hasAccess) {
+            $em = $this->getDoctrine()->getManager();
 
-        return $this->render('cinema_hall/index.html.twig', array(
-            'cinema_halls' => $cinema_halls,
-        ));
+            $cinema_halls = $em->getRepository('AppBundle:Cinema_hall')->findAll();
+
+            return $this->render('cinema_hall/index.html.twig', array(
+                'cinema_halls' => $cinema_halls,
+            ));
+        } else {
+            return $this->redirectToRoute('homepage');
+        }
     }
 
     /**
@@ -39,22 +46,29 @@ class Cinema_hallController extends Controller
      */
     public function newAction(Request $request)
     {
-        $cinema_hall = new Cinema_hall();
-        $form = $this->createForm('AppBundle\Form\Cinema_hallType', $cinema_hall);
-        $form->handleRequest($request);
+        $loggedUserRole = $this->get('security.token_storage')->getToken()->getUser();
+        $hasAccess = $this->hasAccess($loggedUserRole);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($cinema_hall);
-            $em->flush();
+        if($hasAccess) {
+            $cinema_hall = new Cinema_hall();
+            $form = $this->createForm('AppBundle\Form\Cinema_hallType', $cinema_hall);
+            $form->handleRequest($request);
 
-            return $this->redirectToRoute('cinema_hall_show', array('id' => $cinema_hall->getId()));
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->persist($cinema_hall);
+                $em->flush();
+
+                return $this->redirectToRoute('cinema_hall_show', array('id' => $cinema_hall->getId()));
+            }
+
+            return $this->render('cinema_hall/new.html.twig', array(
+                'cinema_hall' => $cinema_hall,
+                'form' => $form->createView(),
+            ));
+        } else {
+            return $this->redirectToRoute('homepage');
         }
-
-        return $this->render('cinema_hall/new.html.twig', array(
-            'cinema_hall' => $cinema_hall,
-            'form' => $form->createView(),
-        ));
     }
 
     /**
@@ -65,12 +79,19 @@ class Cinema_hallController extends Controller
      */
     public function showAction(Cinema_hall $cinema_hall)
     {
-        $deleteForm = $this->createDeleteForm($cinema_hall);
+        $loggedUserRole = $this->get('security.token_storage')->getToken()->getUser();
+        $hasAccess = $this->hasAccess($loggedUserRole);
 
-        return $this->render('cinema_hall/show.html.twig', array(
-            'cinema_hall' => $cinema_hall,
-            'delete_form' => $deleteForm->createView(),
-        ));
+        if($hasAccess) {
+            $deleteForm = $this->createDeleteForm($cinema_hall);
+
+            return $this->render('cinema_hall/show.html.twig', array(
+                'cinema_hall' => $cinema_hall,
+                'delete_form' => $deleteForm->createView(),
+            ));
+        } else {
+            return $this->redirectToRoute('homepage');
+        }
     }
 
     /**
@@ -81,21 +102,28 @@ class Cinema_hallController extends Controller
      */
     public function editAction(Request $request, Cinema_hall $cinema_hall)
     {
-        $deleteForm = $this->createDeleteForm($cinema_hall);
-        $editForm = $this->createForm('AppBundle\Form\Cinema_hallType', $cinema_hall);
-        $editForm->handleRequest($request);
+        $loggedUserRole = $this->get('security.token_storage')->getToken()->getUser();
+        $hasAccess = $this->hasAccess($loggedUserRole);
 
-        if ($editForm->isSubmitted() && $editForm->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+        if($hasAccess) {
+            $deleteForm = $this->createDeleteForm($cinema_hall);
+            $editForm = $this->createForm('AppBundle\Form\Cinema_hallType', $cinema_hall);
+            $editForm->handleRequest($request);
 
-            return $this->redirectToRoute('cinema_hall_edit', array('id' => $cinema_hall->getId()));
+            if ($editForm->isSubmitted() && $editForm->isValid()) {
+                $this->getDoctrine()->getManager()->flush();
+
+                return $this->redirectToRoute('cinema_hall_edit', array('id' => $cinema_hall->getId()));
+            }
+
+            return $this->render('cinema_hall/edit.html.twig', array(
+                'cinema_hall' => $cinema_hall,
+                'edit_form' => $editForm->createView(),
+                'delete_form' => $deleteForm->createView(),
+            ));
+        } else {
+            return $this->redirectToRoute('homepage');
         }
-
-        return $this->render('cinema_hall/edit.html.twig', array(
-            'cinema_hall' => $cinema_hall,
-            'edit_form' => $editForm->createView(),
-            'delete_form' => $deleteForm->createView(),
-        ));
     }
 
     /**
@@ -106,16 +134,23 @@ class Cinema_hallController extends Controller
      */
     public function deleteAction(Request $request, Cinema_hall $cinema_hall)
     {
-        $form = $this->createDeleteForm($cinema_hall);
-        $form->handleRequest($request);
+        $loggedUserRole = $this->get('security.token_storage')->getToken()->getUser();
+        $hasAccess = $this->hasAccess($loggedUserRole);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($cinema_hall);
-            $em->flush();
+        if($hasAccess) {
+            $form = $this->createDeleteForm($cinema_hall);
+            $form->handleRequest($request);
+
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($cinema_hall);
+                $em->flush();
+            }
+
+            return $this->redirectToRoute('cinema_hall_index');
+        } else {
+            return $this->redirectToRoute('homepage');
         }
-
-        return $this->redirectToRoute('cinema_hall_index');
     }
 
     /**
@@ -132,5 +167,18 @@ class Cinema_hallController extends Controller
             ->setMethod('DELETE')
             ->getForm()
         ;
+    }
+
+    private function hasAccess($loggedUserRole)
+    {
+        if($loggedUserRole !== 'anon.') {
+            if($loggedUserRole->getRole() === 'admin') {
+                return true;
+            } else {
+                return false;
+            }
+        } else {
+            return false;
+        }
     }
 }
